+++ game/data/Constants.java
@@ -1,0 +1,17 @@
+package game.data;
+
+public class Constants {
+
+    public Constants() {}
+
+    public static final int NO_OF_STATES = 37;
+    public static final double VAT = 0.075;
+    public static final double MINIMUM_CASH = 600000.0;
+    public static final long MINIMUM_POPULATION = 25000;
+    public static final double MINIMUM_GDP = 4000000.0;
+    public static final long INFLATION_CONSTANT = 300000;
+    public static final long BASE_OIL = 46000; //Nigeria's average oil production at 46M barrels per month. // **IMPORTANT** This has been scaled down 
+    public static final int REPORT_DELAY_TIME = 25;
+    
+
+}