+++ utils/Sleeper.java
@@ -2,6 +2,7 @@ package utils;
 import java.util.*;
 
 public class Sleeper {
+    //Take a number in milliseconds and stop the thread for that amount of time
     public static void sleep(long millis) {
         try {
             Thread.sleep(millis);