+++ utils/StateSelecter.java
@@ -2,6 +2,7 @@ package utils;
 import java.util.*;
 
 public class StateSelecter {
+    //Take a state as input and set it as the user's chosen state
     public static String stateSelecter(String selectedState) {
         String[] availableStates = {"Abia", "Adamawa", "Akwa Ibom", "Anambra", "Bauchi", "Bayelsa", "Benue", "Borno", "Cross River", "Delta", "Ebonyi",
         "Edo", "Ekiti", "Enugu", "Gombe", "Imo", "Jigawa", "Kaduna", "Kano", "Katsina", "Kebbi", "Kogi", "Kwara", "Lagos", "Nasarawa", "Niger", "Ogun",